refactor(requirements): drop utils dir, move helpers to src root
Address Sanderson review: avoid the "utils" naming and the nested
src/requirements/utils path (src/requirements is reserved for actual
requirement rules). Move workspaces.ts and stripComments.ts directly
under packages/requirements/src/.
