+++ packages/requirements/src/requirements/public-package-dependencies/requireConnectPublicDependencies.ts
@@ -1,8 +1,8 @@
 import { mkdirSync, readFileSync, writeFileSync } from 'node:fs';
 import { join } from 'node:path';
 
+import { listAllWorkspaces, readPackageJson } from '../../workspaces';
 import type { Requirement } from '../Requirement';
-import { listAllWorkspaces, readPackageJson } from '../utils/workspaces';
 
 type PackageJson = {
     readonly name?: string;
