+++ packages/requirements/src/stripComments.ts
